product, customer and company has been updated. File names are changed.
