--- 05.05.2021/CRMSystem.java
@@ -1,15 +1,0 @@
-
-public class CRMSystem {
-
-    public User login(String userID){return null;}
-
-
-    private void showMenu(User user){}
-    private void adminMenu(){}
-    private void developerMenu(){}
-    private void customerMenu(){}
-
-
-    public boolean singUp(String name, String surName, String email, String phoneNumber){return true;}
-
-}
