--- 05.05.2021/Schedule.java
@@ -1,13 +1,0 @@
-import java.util.Date;
-
-public class Schedule implements Comparable<Schedule>{
-    private Date date;
-    private String process;
-
-    public Schedule(Date date, String process){}
-
-    @Override
-    public int compareTo(Schedule o) {
-        return 0;
-    }
-}
