--- 05.05.2021/User.java
@@ -1,31 +1,0 @@
-import java.util.Date;
-import java.util.PriorityQueue;
-
-public class User implements UserInterface{
-    private String name;
-    private String surName;
-    private String ID;
-    private String password;
-
-
-
-    public String getName(){
-        return name;
-    }
-
-    public String getSurName(){
-        return surName;
-    }
-
-    public String getID(){
-        return ID;
-    }
-
-    public String getPassword(){
-        return password;
-    }
-
-    public void setName(){}
-    public void setSurName(){}
-    public void setPassword(){};
-}
