--- 05.05.2021/UserInformationSystem.java
@@ -1,4 +1,0 @@
-public class UserInformationSystem {
-    private String message;
-    private User user;
-}
