--- 05.05.2021/UserInterface.java
@@ -1,10 +1,0 @@
-public interface UserInterface {
-    String getName();
-    String getSurName();
-    String getID();
-    String getPassword();
-
-    void setName();
-    void setSurName();
-    void setPassword();
-}
