+++ src/UserInformationSystem.java
@@ -1,0 +1,5 @@
+package src;
+public class UserInformationSystem {
+    private String message;
+    private User user;
+}
