+++ src/UserInterface.java
@@ -1,0 +1,16 @@
+package src;
+public interface UserInterface {
+    public String getName();
+
+    public String getSurName();
+
+    public String getID();
+
+    public String getPassword();
+
+    void setName(String name);
+
+    void setSurName(String surName);
+
+    void setPassword(String password);
+}
